Use 32 bits for locations and lengths
* Note that this changes requires `bundle exec rake clean && bundle exec rake clean`.
  This is because the C extension depends on yarp.h and other headers,
  but the Rake::ExtensionTask does not understand that dependency.
  The :compile task runs but does not recompile since it sees no change in ext/yarp.
  This is addressed in the next commit.
* 64 bits is quite wasteful in the serialization format
  and Ruby files should never be larger than 4GB.
* As an example `YARP.dump("1+2").bytesize` goes from 270 to 146 bytes.

+++ src/error.h
@@ -29,7 +29,7 @@ yp_error_list_init(yp_error_list_t *error_list);
 
 // Append an error to the given list of errors.
 void
-yp_error_list_append(yp_error_list_t *error_list, const char *message, uint64_t position);
+yp_error_list_append(yp_error_list_t *error_list, const char *message, uint32_t position);
 
 // Deallocate the internal state of the given error list.
 void
